Fix unsigned vs signed compare.
git-svn-id: file:///home/jj/hercules.svn/trunk@6212 956126f8-22a0-4046-8f4a-272fa8102e63

+++ cmdtab.c
@@ -326,7 +326,7 @@ int HelpCommand(int argc, char *argv[], char *cmdline)
 		            WRMSG ( HHC01603, "I", "Long description:");
                     while(pCmdTab->longdesc[i])
                     {
-                        for(j = 0; j < sizeof(buf)      && 
+                        for(j = 0; j < (int)sizeof(buf)      && 
                                    pCmdTab->longdesc[i] && 
                                    pCmdTab->longdesc[i] != '\n'; i++, j++)
                         {
